Explaining how to package ReactNative for Nexus/Maven deployment.
Summary:
**Motivation**

I needed to do this today. It's explained in code in the repository, but not included in the actual documentation. As React Native continues to grow, this is something that will be used more and more by the community.

**Test plan (required)**

No test plan required, this is simply a documentation change to add things that are only explained in code comments currently.
Closes https://github.com/facebook/react-native/pull/12938

Differential Revision: D4709339

Pulled By: mkonicek

fbshipit-source-id: a56c7573c3fa25a4059657b95b482b641ff229ff