Shorten template instructions.
Summary:
I've noticed many pull requests are opened with a very short description that fails to explain the motivation behind the PR, and many more forego providing a test plan. With this PR, I am shortening the template in order to provide concise instructions that provide the essential steps above the fold in the "Open a pull request" composer window.

- We need people to open a PR against master, not stable. The exact reason for this is not important and providing an explanation takes away from other more important points.
- Test plans are essential, but their requirement appears below the fold in the current template.
- More PRs could use tests.
- Make a point of asking PR authors to follow up on CI test failures.
- The composer does not parse Markdown into HTML. Markdown is pretty readable as is, but using reference links instead of inline links should help with readability.

I observed that it will only display the first 8 lines of the PR template above the fold. Seeing t
Closes https://github.com/facebook/react-native/pull/12958

Differential Revision: D4718294

Pulled By: mkonicek

fbshipit-source-id: b19a29e5ed73fb78d09c7de17625b1883590075c