Remove old examples from RN repo
Summary:
React Native originally shipped with these examples, which have gotten super old and unmaintained.

Cruft like this in the core repo only slows us down with extra issues and pull requests. Eventually we may create a different repo for samples that will be maintained into the future.

We are doing some spring cleaning in the repo, so now is a good time to remove these and let the community share their own examples.
Closes https://github.com/facebook/react-native/pull/13064

Differential Revision: D4770699

Pulled By: ericvicenti

fbshipit-source-id: bbcdec336b28e5a9330afbdb6b166e94e15fa485